Deploy uniswap/ens backed contracts (#13)
This sets up a deploy function, `deployAll` that deploys the bytecode of ENS and Uniswap, sets up proper liquidity pairs for Uniswap and hooks everything up with our contract. This allows us to test everything using the "real" contracts, instead of the dummy ones.

As you'll notice, the registration fee taken from the Uniswap price oracle is always going to be slightly lower than expected, eg. given a 1:1 liquidity ratio, the price of 1000 USD will be ~999 ETH.